+++ CHANGELOG.md
@@ -81,6 +81,7 @@
 * Add guice bindings report (`printGuiceBindings()` and `printAllGuiceBindings()` methods on main bundle)
 * (breaking) `DebugGuiceyLifecycle` listener renamed into `LifecycleDiagnostic`
 * (breaking) All diagnostic reports moved into one top-level "debug" package.
+* Add guice aop appliance report
 
 Main breaking changes were caused by:
  - jersey 2.26 introduces an abstraction for injection layer in order to get rid of hk2 direct usage.